Fix: Prevent double destroy of Layout causing Yoga WASM memory errors (#117)
* refactor: enhance styling capabilities with alignSelf support

- Updated the YogaStyles interface to include AlignSelf type.
- Added alignSelf mapping in applyStyle.ts for improved style application.
- Introduced new Tailwind utility for 'self-auto' alignment.
- Updated .gitignore to include *.tgz files.

* refactor: optimize child sorting logic in onChildAdded function

Replaces direct access to pixiParent.children with a local variable for improved readability and performance. This change simplifies the logic for determining the last child and enhances maintainability of the code.

* Changes

* feat: add destroyed flag to Layout class for resource management

Introduces a 'destroyed' property to the Layout class to track the object's state. Updates the destroy method to prevent multiple calls and ensure proper resource cleanup when the object is destroyed.

* Revert to pixijs package.json